Rename TokenRepository
